Fixes issue #2999 - Add JUnit tests for all of the WebApplicationResponse API
